moved helper fn to utils folder for cleaner code
